Fix bugs found in a second correctness sweep
Crashes (previously unreachable features):
- warmup_epochs > 0 raised AttributeError on the first training step:
  PatchTransformer is a plain nn.Module, so Lightning never gave it current_epoch.
  The parent now propagates it; it stays out of state_dict, so checkpoints are
  unaffected. All 98 historical configs set -1, so no completed run is affected.
- backbone="unet" raised ModuleNotFoundError from a pre-rename import path
  (`from models.DenseNet import ...`; there is no top-level `models` package).
- spconv was documented as optional but imported at module scope from a module
  reachable by every entry point, so a default install could not import the
  package at all. Deferred into the two sparse backbones that need it.

Wrong results:
- SurfaceDistanceLoss returned torch.norm(mean) rather than a mean, inflating the
  term by ~sqrt(n_landmarks) (~5.9x at 35 landmarks) whenever "SD" was compounded
  with other losses. It also ignored the loss mask.
- PatchExtractor truncated patch centres with .long() instead of rounding, biasing
  every patch up to a voxel toward the origin.

Rejected valid input:
- CompoundLoss asserted sum(weights) == 1.0 exactly, so [0.7, 0.2, 0.1] raised;
  meanwhile its own default was [1.0]*n, summing to n. Now a 1e-6 tolerance, a
  ValueError rather than a stripped assert, and a 1/n default.
- --set_zoom parsed with int, so 0.8mm spacing could not be requested at all.

Also closes a guard gap (use_coarse_pred=False silently dropped the coarse
features it still received) and corrects infer.py's module docstring, which
stated the revert order backwards relative to the code it links to.

104 tests pass. Behaviour changes recorded in CHANGES.md.

+++ src/verpex/cli/prepare_data.py
@@ -536,7 +536,7 @@ def main() -> None:
 
     parser.add_argument(
         "--set_zoom",
-        type=lambda x: tuple(map(int, x.split(","))),
+        type=lambda x: tuple(map(float, x.split(","))),
         help="Zoom for rescaling (format: x,y,z)",
         default=(1, 1, 1),
     )
